refactor: removed dependencies from dto
