+++ Source/SeaInk.Infrastructure/SeaInk.Infrastructure.Dto/StudyAssignmentDto.cs
@@ -1,0 +1,12 @@
+namespace SeaInk.Infrastructure.Dto
+{
+    public record StudyAssignmentDto(
+        Guid Id,
+        int UniversityId,
+        string Title,
+        bool IsMilestone,
+        DateTime StartDate,
+        DateTime EndDate,
+        double MinPoints,
+        double MaxPoints);
+}
