+++ Source/SeaInk.Infrastructure/SeaInk.Infrastructure.Dto/StudyStudentGroupDto.cs
@@ -1,0 +1,3 @@
+namespace SeaInk.Infrastructure.Dto;
+
+public record StudyStudentGroupDto(Guid Id, string GroupName);
