+++ Source/SeaInk.Infrastructure/SeaInk.Infrastructure.Dto/SubjectDto.cs
@@ -1,0 +1,3 @@
+namespace SeaInk.Infrastructure.Dto;
+
+public record SubjectDto(Guid Id, int UniversityId, string Name, IReadOnlyCollection<StudyAssignmentDto> Assignments);
